extract Keydetails into a a common dependency
